chore(api,lib): move young schema to lib
