+++ api/src/models/index.ts
@@ -23,7 +23,8 @@ export { SessionPhase1TokenModel } from "./sessionPhase1Token";
 export { StructureModel, StructureDocument, StructureType } from "./structure";
 export { TagsModel, TagsDocument, TagsType } from "./tags";
 export { WaitingListModel, WaitingListDocument, WaitingListType } from "./waitingList";
-export { YoungModel, YoungDocument, YoungType } from "./young";
+export { YoungModel, YoungDocument } from "./young";
+export { YoungType } from "snu-lib";
 
 export { ClasseModel, ClasseDocument } from "./cle/classe";
 export { EtablissementModel, EtablissementDocument } from "./cle/etablissement";
