+++ packages/lib/src/mongoSchema/index.ts
@@ -5,3 +5,4 @@ export * from "./cle/etablissement";
 export * from "./cohesionCenter";
 export * from "./pointDeRassemblement";
 export * from "./referent";
+export * from "./young";
